fix compile explicit casts, 'new' expressions can be simplified, refactor spells

+++ Source/Services/Mangos.World/Handlers/WS_Handlers_Chat.cs
@@ -38,7 +38,7 @@ namespace Mangos.World.Handlers
             {
                 return 1;
             }
-            return objCharacter.DND ? 2 : 0;
+            return (byte)(objCharacter.DND ? 2 : 0);
         }
 
         public void On_CMSG_MESSAGECHAT(ref Packets.PacketClass packet, ref WS_Network.ClientClass client)
